scoped type checking now works.
closes #13
