+++ src/main/java/parser/BioScriptParser.java
@@ -24,14 +24,12 @@ public class BioScriptParser implements Phase {
     public static final Logger logger = LogManager.getLogger(BioScriptParser.class);
     private BSParser parser;
     private InferenceConfig config = ConfigFactory.getConfig();
-    private BSTypeChecker typeChecker;
     private BSSymbolTable symbolTable;
     private String file;
 
     public BioScriptParser(String fileName) {
         this.file = fileName;
         this.symbolTable = new BSSymbolTable();
-        this.typeChecker = new BSTypeChecker(this.symbolTable.getSymbolTable());
     }
 
     @Override
@@ -46,10 +44,8 @@ public class BioScriptParser implements Phase {
             try {
                 BSProgram program = this.parser.BSProgram();
                 program.accept(this.symbolTable);
-                // logger.info(this.symbolTable);
                 if (!this.config.getErrorLevel().disabled()) {
-                    program.accept(this.typeChecker);
-                    this.typeChecker.solve();
+                    this.symbolTable.solve();
                 } else {
                     logger.error("Type checking has been disabled.");
                 }
