+++ src/main/java/typesystem/rules/Rule.java
@@ -13,7 +13,7 @@ import java.util.Set;
 
 import config.ConfigFactory;
 import config.InferenceConfig;
-import shared.variables.Variable;
+import shared.Variable;
 import typesystem.Inference.InferenceType;
 import typesystem.elements.Formula;
 import chemical.epa.ChemTypes;
